// react-module/src/openapi/models/atomic-operation-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicOperationBasicDto
 */
export interface AtomicOperationBasicDto {
    /**
     * 
     * @type {string}
     * @memberof AtomicOperationBasicDto
     */
    operationResult?: AtomicOperationBasicDtoOperationResultEnum;
    /**
     * 
     * @type {number}
     * @memberof AtomicOperationBasicDto
     */
    id?: any;
    /**
     * 
     * @type {OperationTypeBasicDto}
     * @memberof AtomicOperationBasicDto
     */
    operationType?: any;
    /**
     * 
     * @type {number}
     * @memberof AtomicOperationBasicDto
     */
    priority?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductBasicDto&gt;}
     * @memberof AtomicOperationBasicDto
     */
    inputProducts?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductBasicDto&gt;}
     * @memberof AtomicOperationBasicDto
     */
    outputProducts?: any;
    /**
     * 
     * @type {ProductionOrderBasicDto}
     * @memberof AtomicOperationBasicDto
     */
    productionOrder?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicOperationBasicDto
     */
    groupingOption?: AtomicOperationBasicDtoGroupingOptionEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicOperationBasicDtoOperationResultEnum {
    SUCCESS = 'SUCCESS',
    BROKEN = 'BROKEN',
    ASSIGNED = 'ASSIGNED',
    NOTASSIGNED = 'NOT_ASSIGNED',
    ISRUNNING = 'IS_RUNNING',
    COMPLETED = 'COMPLETED',
    POSTPONED = 'POSTPONED'
}
/**
    * @export
    * @enum {string}
    */
export enum AtomicOperationBasicDtoGroupingOptionEnum {
    EACHOPERATIONTOSEPARATECYCLE = 'EACH_OPERATION_TO_SEPARATE_CYCLE',
    ALLOPERATIONTOONECYCLE = 'ALL_OPERATION_TO_ONE_CYCLE'
}